Search fixes + small evaluation change + make/take move speed up
Fixed the aspiration window and the PV line output in the search.
Added a penalty for having the king on an open file and adjusted the mobility weights, also removed the penatly for the king not being able to castle while being on a weak square becuase the bot seemed to perform better against other bots without it.
Changed take move functions to change the side manually instead of getting it from the undo info, also make and take move will adjust the big pieces counter manually. This removes some information from the undo info and speeds up the move handling a small but noticeable amount during perft.
Added a check for big pieces in null move pruning to (hopefully) not do null move pruning in zugzwang.